// src/database/sp/interfaces/insercion-incidencias-params.sp.ts
export interface IInsercionIncidenciasSpParams {
    codigoAduana: string;
    codigoRegimen: string;
    anio: string;
    orden: string;
    incidencia: string;
    tipoIncidencia: string;
    codigoIncidencia: string;
    ctrlTrans: string;
    persReci: string;
    agente: string;
    marca: string;
    ctrlImp: string;
    observacion: string;
    subIncidencia: string;
    nombreSubIncidencia: string;
    fechaProg: string;
    flag: string;
    pcIp: string;
    ippc: string;
    flag2: string;
    userId: number;
}